+++ test/inc_test.dart
@@ -4,27 +4,23 @@ import 'package:test/test.dart';
 void main() {
   group('Inc', () {
     test('Compile test string', () async {
-      final inc = Inc(
-        input: "Hello world, |>print('tester', end='')<|, thanks!",
-      );
+      final inc = Inc("Hello world, |>print('tester', end='')<|, thanks!");
 
       final result = await inc.compile();
       expect(result, 'Hello world, tester, thanks!');
     });
 
     test('Compile nested test strings', () async {
-      final inc = Inc(
-        input: "|>print('outer with |>print('inner', end='')<|', end='')<|",
-      );
+      final inc =
+          Inc("|>print('outer with |>print('inner', end='')<|', end='')<|");
 
       final result = await inc.compile();
       expect(result, 'outer with inner');
     });
 
     test('Compile with adjacent test strings', () async {
-      final inc = Inc(
-        input: "|>print('first', end='')<| then |>print('next', end='')<|",
-      );
+      final inc =
+          Inc("|>print('first', end='')<| then |>print('next', end='')<|");
 
       final result = await inc.compile();
       expect(result, 'first then next');
@@ -32,7 +28,7 @@ void main() {
 
     test('Compile with multiline test string', () async {
       final inc = Inc(
-        input: """
+        """
 |>
 x = 10
 y = 5
@@ -45,45 +41,35 @@ print(x + y, end='')
     });
 
     test('Compile with a regular string', () async {
-      final inc = Inc(
-        input: 'regular string here!',
-      );
+      final inc = Inc('regular string here!');
 
       final result = await inc.compile();
       expect(result, 'regular string here!');
     });
 
     test('Compile with no beginning pattern', () async {
-      final inc = Inc(
-        input: 'here is a non<| compile',
-      );
+      final inc = Inc('here is a non<| compile');
 
       final result = await inc.compile();
       expect(result, 'here is a non<| compile');
     });
 
     test('Compile with no ending pattern', () async {
-      final inc = Inc(
-        input: 'here |>is a non compile',
-      );
+      final inc = Inc('here |>is a non compile');
 
       final result = await inc.compile();
       expect(result, 'here |>is a non compile');
     });
 
     test('Compile with dangling beginning pattern', () async {
-      final inc = Inc(
-        input: "here |>|>print('only once', end='')<|",
-      );
+      final inc = Inc("here |>|>print('only once', end='')<|");
 
       final result = await inc.compile();
       expect(result, 'here |>only once');
     });
 
     test('Compile with dangling ending pattern', () async {
-      final inc = Inc(
-        input: "here |>print('only once', end='')<|<|",
-      );
+      final inc = Inc("here |>print('only once', end='')<|<|");
 
       final result = await inc.compile();
       expect(result, 'here only once<|');
@@ -91,7 +77,7 @@ print(x + y, end='')
 
     test('Compile with non-default patterns', () async {
       final inc = Inc(
-        input: "hello [print('world', end='')]",
+        "hello [print('world', end='')]",
         beginPattern: '[',
         endPattern: ']',
       );
@@ -102,7 +88,7 @@ print(x + y, end='')
 
     test('Compile with non-default identical patterns', () async {
       final inc = Inc(
-        input: "hello *print('world', end='')*",
+        "hello *print('world', end='')*",
         beginPattern: '*',
         endPattern: '*',
       );
@@ -113,7 +99,7 @@ print(x + y, end='')
 
     test('Compile with adjacent non-default identical patterns', () async {
       final inc = Inc(
-        input: "*print('first', end='')* then *print('next', end='')*",
+        "*print('first', end='')* then *print('next', end='')*",
         beginPattern: '*',
         endPattern: '*',
       );