+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+* Use positional parameters when applicable
+* Merge evaluator class with inc class
+
 ## 1.0.1+1
 
 * Fix example